infrastructure/lightsail/requirements.txt
# Minimal requirements for Lightsail streaming service

# Core dependencies
schwab-py>=1.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
structlog>=23.0.0
pandas>=2.0.0

# WebSocket support
websocket-client>=1.6.0

# Data handling
python-dateutil>=2.8.0

# Pydantic for data models
pydantic>=2.0.0

